add HiDPI texture/window ratio for MacOS
This patch creates the ```SLDUI_Core.texture_window_hdpi_ratio_``` values initialised in ```SDLUI_Init```.
Then mutliplies all MouseState x&y by those factors to get to mouse on the correct place on HiDPI screens (for  Mac only)
It's not checking wether the window moves from a monitor to another though (I don't know how to do that). So no updates will be made if the ratio changes when moving the window to a different Monitor.